pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "costlocker-cli"
version = "0.1.0"
description = "Sync Google Calendar events to Costlocker timesheets"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "rich>=13",
    "httpx>=0.27",
    "google-api-python-client>=2.130",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.2",
    "pendulum>=3",
]

[project.scripts]
costlocker = "costlocker_cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["costlocker_cli*"]
